execve & path

+++ libft/srcs/strings/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: odroz-ba <odroz-ba@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 13:48:16 by odroz-ba          #+#    #+#             */
-/*   Updated: 2021/04/23 17:28:39 by odroz-ba         ###   ########lyon.fr   */
+/*   Updated: 2021/05/19 16:25:32 by odroz-ba         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
